# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Emergence World simulation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Position {
  x: number;
  z: number;
}

export interface Agent {
  id: string;
  name: string;
  role: string;
  location: string;
  position: Position;
  mood: string;
  energy: number;
  credits: number;
  isCurrentTurn: boolean;
  lastAction?: string;
  /** @nullable */
  lastSpeech?: string | null;
  relationships?: string[];
  animation: string;
  color: string;
  portrait?: string;
}

export interface AgentDetail {
  id: string;
  name: string;
  role: string;
  personality: string;
  northStar: string;
  location: string;
  position: Position;
  mood: string;
  energy: number;
  knowledge: number;
  influence: number;
  credits: number;
  isCurrentTurn: boolean;
  lastAction?: string;
  /** @nullable */
  lastSpeech?: string | null;
  color: string;
  portrait?: string;
}

export interface Landmark {
  id: string;
  name: string;
  category: string;
  capacity: number;
  description: string;
  position: Position;
  occupants: string[];
  tools: string[];
}

export interface Weather {
  condition: string;
  temperature: number;
  description: string;
  icon?: string;
  worldTime: string;
  timezone: string;
}

export interface WorldState {
  agents: Agent[];
  weather: Weather;
  worldTime: string;
  /** @nullable */
  currentTurnAgent: string | null;
  turnNumber: number;
  dayNumber: number;
  isRunning: boolean;
  totalAgents: number;
}

export interface SimulationStatus {
  isRunning: boolean;
  /** @nullable */
  currentTurnAgent?: string | null;
  turnNumber: number;
  dayNumber: number;
  message?: string;
}

export interface Memory {
  id: string;
  agentId: string;
  content: string;
  type: string;
  timestamp: string;
  /** @nullable */
  location?: string | null;
}

export interface Conversation {
  id: string;
  speakerName: string;
  speakerId: string;
  message: string;
  location: string;
  timestamp: string;
  type: string;
  nearbyAgents: string[];
}

export interface CreditBalance {
  agentId: string;
  agentName: string;
  balance: number;
  earned: number;
  spent: number;
}

export interface Transaction {
  id: string;
  /** @nullable */
  fromAgent?: string | null;
  toAgent: string;
  amount: number;
  reason: string;
  timestamp: string;
}

export interface ConstitutionArticle {
  number: number;
  title: string;
  content: string;
  /** @nullable */
  amendedBy?: string | null;
  /** @nullable */
  amendedAt?: string | null;
}

export interface Constitution {
  articles: ConstitutionArticle[];
  version: number;
  lastAmended?: string;
}

export type ProposalVotesItem = {
  agentId: string;
  vote: string;
  reason?: string;
};

export interface Proposal {
  id: string;
  title: string;
  description: string;
  proposedBy: string;
  status: string;
  votesFor: number;
  votesAgainst: number;
  totalVoters: number;
  timestamp: string;
  votes: ProposalVotesItem[];
}

export interface BlogPost {
  id: string;
  authorId: string;
  authorName: string;
  title: string;
  content: string;
  timestamp: string;
  tags: string[];
}

export interface BillboardPost {
  id: string;
  authorId: string;
  authorName: string;
  message: string;
  timestamp: string;
}

export type ActivityEventMetadata = { [key: string]: unknown };

export interface ActivityEvent {
  id: string;
  agentId: string;
  agentName: string;
  type: string;
  description: string;
  location: string;
  timestamp: string;
  metadata?: ActivityEventMetadata;
}

export interface AWIMetrics {
  m1_population: number;
  m2_safety_score: number;
  m3_exploration: number;
  m4_tool_usage: number;
  m5_governance: number;
  m6_expression: number;
  m7_social_fabric: number;
  m8_economic_vitality: number;
  m9_constitutional_growth: number;
  dayNumber: number;
}

export type GetConversationsParams = {
limit?: number;
locationId?: string;
};

export type GetTransactionsParams = {
limit?: number;
};

export type GetBlogsParams = {
agentId?: string;
};

export type GetRecentActivityParams = {
limit?: number;
};