# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Emergence World simulation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get a specific agent's full profile and state
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDetail"
        "404":
          description: Agent not found
  /agents/{agentId}/memories:
    get:
      operationId: getAgentMemories
      tags: [agents]
      summary: Get an agent's recent memories
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent memories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
...
components:
  schemas:
...
    Position:
      type: object
      properties:
        x:
          type: number
        z:
          type: number
      required: [x, z]
...
    AgentDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        personality:
          type: string
        northStar:
          type: string
        location:
          type: string
        position:
          $ref: "#/components/schemas/Position"
        mood:
          type: string
        energy:
          type: number
        knowledge:
          type: number
        influence:
          type: number
        credits:
          type: number
        isCurrentTurn:
          type: boolean
        lastAction:
          type: string
        lastSpeech:
          type: ["string", "null"]
        color:
          type: string
        portrait:
          type: string
      required: [id, name, role, personality, northStar, location, position, mood, energy, knowledge, influence, credits, isCurrentTurn, color]
...
    Memory:
      type: object
      properties:
        id:
          type: string
        agentId:
          type: string
        content:
          type: string
        type:
          type: string
        timestamp:
          type: string
        location:
          type: ["string", "null"]
      required: [id, agentId, content, type, timestamp]
...