# lib/api-zod/src/generated/types/agentDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Emergence World simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { Position } from './position';

export interface AgentDetail {
  id: string;
  name: string;
  role: string;
  personality: string;
  northStar: string;
  location: string;
  position: Position;
  mood: string;
  energy: number;
  knowledge: number;
  influence: number;
  credits: number;
  isCurrentTurn: boolean;
  lastAction?: string;
  /** @nullable */
  lastSpeech?: string | null;
  color: string;
  portrait?: string;
}